Split unit tests to separate files
